Clarify integer types handling
Reduce amount of function in tools.
