+++ src/metadata/resolution.cc
@@ -34,7 +34,7 @@ Resolution::Resolution(const std::string& string)
 
     if (!parts[0].empty()) {
         auto x = stoulString(trimString(parts[0]));
-        if (x == 0 || x == std::numeric_limits<uint64_t>::max()) {
+        if (x == 0 || x == std::numeric_limits<unsigned long>::max()) {
             throw_std_runtime_error("Failed to parse '{}' to valid resolution", string);
         }
         _x = x;
@@ -42,7 +42,7 @@ Resolution::Resolution(const std::string& string)
 
     if (!parts[1].empty()) {
         auto y = stoulString(trimString(parts[1]));
-        if (y == 0 || y == std::numeric_limits<uint64_t>::max()) {
+        if (y == 0 || y == std::numeric_limits<unsigned long>::max()) {
             throw_std_runtime_error("Failed to parse '{}' to valid resolution", string);
         }
         _y = y;
@@ -53,10 +53,10 @@ Resolution::Resolution(std::uint64_t x, std::uint64_t y)
     : _x(x)
     , _y(y)
 {
-    if (x == 0 || x == std::numeric_limits<uint64_t>::max()) {
+    if (x == 0 || x == std::numeric_limits<std::uint64_t>::max()) {
         throw_std_runtime_error("Failed to set '{}x{}' as resolution", x, y);
     }
-    if (y == 0 || y == std::numeric_limits<uint64_t>::max()) {
+    if (y == 0 || y == std::numeric_limits<std::uint64_t>::max()) {
         throw_std_runtime_error("Failed to set '{}x{}' as resolution", x, y);
     }
 }
