+++ src/util/grb_net.cc
@@ -215,3 +215,10 @@ std::string GrbNet::ipToInterface(std::string_view ip)
     log_warning("Failed to find interface for IP: {}", ip);
     return {};
 }
+
+std::string GrbNet::renderWebUri(std::string_view ip, in_port_t port)
+{
+    if (ip.find(':') != std::string_view::npos)
+        return fmt::format("[{}]:{}", ip, port);
+    return fmt::format("{}:{}", ip, port);
+}
