+++ src/util/url_utils.cc
@@ -21,12 +21,13 @@ Gerbera - https://gerbera.io/
 
 /// \file url_utils.cc
 
+#include "url_utils.h" // API
+
+#include <fmt/format.h>
+#include <sstream>
+
 #include "common.h"
 #include "exceptions.h"
-#include "tools.h"
-#include <fmt/format.h>
-
-#include "url_utils.h"
 
 namespace URLUtils {
 
@@ -71,6 +72,147 @@ std::map<std::string, std::string> parseParameters(std::string_view filename, st
 {
     const auto parameters = filename.substr(baseLink.size());
     return pathToMap(parameters);
+}
+
+static constexpr auto hexCharS2 = "0123456789ABCDEF";
+
+/// \brief Converts a string to a URL (meaning: %20 instead of space and so on)
+/// \param str String to be converted.
+/// \return string that contains the url-escaped representation of the original string.
+std::string urlEscape(std::string_view str)
+{
+    std::ostringstream buf;
+    for (std::size_t i = 0; i < str.length();) {
+        auto c = str[i];
+        int cplen = 1;
+        if ((c & 0xf8) == 0xf0)
+            cplen = 4;
+        else if ((c & 0xf0) == 0xe0)
+            cplen = 3;
+        else if ((c & 0xe0) == 0xc0)
+            cplen = 2;
+        if ((i + cplen) > str.length())
+            cplen = 1;
+
+        if ((c >= '0' && c <= '9') || (c >= 'A' && c <= 'Z') || (c >= 'a' && c <= 'z') || c == '_' || c == '-' || c == '.') {
+            buf << char(c);
+        } else {
+            int hi = c >> 4;
+            int lo = c & 15;
+            if (cplen > 1)
+                buf << str.substr(i, cplen);
+            else
+                buf << '%' << hexCharS2[hi] << hexCharS2[lo];
+        }
+        i += cplen;
+    }
+    return buf.str();
+}
+
+/// \brief Opposite of urlEscape :)
+std::string urlUnescape(std::string_view str)
+{
+    auto data = str.data();
+    std::size_t len = str.length();
+    std::ostringstream buf;
+
+    std::size_t i = 0;
+    while (i < len) {
+        const char c = data[i++];
+        if (c == '%') {
+            if (i + 2 > len)
+                break; // avoid buffer overrun
+            char chi = data[i++];
+            char clo = data[i++];
+
+            auto pos = std::strchr(hexCharS2, chi);
+            int hi = pos ? pos - hexCharS2 : 0;
+
+            pos = std::strchr(hexCharS2, clo);
+            int lo = pos ? pos - hexCharS2 : 0;
+
+            int ascii = (hi << 4) | lo;
+            buf << char(ascii);
+        } else if (c == '+') {
+            buf << ' ';
+        } else {
+            buf << c;
+        }
+    }
+    return buf.str();
+}
+
+static std::string dictEncode(const std::map<std::string, std::string>& dict, char sep1, char sep2)
+{
+    std::ostringstream buf;
+    for (auto it = dict.begin(); it != dict.end(); ++it) {
+        if (it != dict.begin())
+            buf << sep1;
+        buf << urlEscape(it->first) << sep2
+            << urlEscape(it->second);
+    }
+    return buf.str();
+}
+
+std::string dictEncode(const std::map<std::string, std::string>& dict)
+{
+    return dictEncode(dict, '&', '=');
+}
+
+std::string dictEncodeSimple(const std::map<std::string, std::string>& dict)
+{
+    return dictEncode(dict, '/', '/');
+}
+
+std::map<std::string, std::string> dictDecode(std::string_view url, bool unEscape)
+{
+    std::map<std::string, std::string> dict;
+    const char* data = url.data();
+    const char* dataEnd = data + url.length();
+    while (data < dataEnd) {
+        const char* ampPos = std::strchr(data, '&');
+        if (!ampPos) {
+            ampPos = dataEnd;
+        }
+        const char* eqPos = std::strchr(data, '=');
+        if (eqPos && eqPos < ampPos) {
+            auto key = std::string_view(data, eqPos - data);
+            auto value = std::string_view(eqPos + 1, ampPos - eqPos - 1);
+            if (unEscape) {
+                dict.try_emplace(urlUnescape(key), urlUnescape(value));
+            } else {
+                dict.emplace(key, value);
+            }
+        }
+        data = ampPos + 1;
+    }
+    return dict;
+}
+
+std::map<std::string, std::string> pathToMap(std::string_view url)
+{
+    std::map<std::string, std::string> out;
+    std::size_t pos;
+    std::size_t lastPos = 0;
+    std::size_t size = url.size();
+    do {
+        pos = url.find('/', lastPos);
+        if (pos == std::string_view::npos)
+            pos = size;
+
+        std::string key = urlUnescape(url.substr(lastPos, pos - lastPos));
+        lastPos = pos == size ? size : pos + 1;
+        pos = url.find('/', lastPos);
+        if (pos == std::string::npos)
+            pos = size;
+
+        std::string value = urlUnescape(url.substr(lastPos, pos - lastPos));
+        lastPos = pos + 1;
+
+        out.emplace(key, value);
+    } while (lastPos < size);
+
+    return out;
 }
 
 } // namespace URLUtils
